drk

Module cli_util

CLI utility functions

Functions

generate_completions

Auxiliary function to generate provided shell completions.

kaching

Fun police go away

parse_token_pair

Auxiliary function to parse provided string into a tokens pair.

parse_tx_from_stdin

Auxiliary function to parse a base64 encoded transaction from stdin.

parse_value_pair

Auxiliary function to parse provided string into a values pair.