drk

Module interactive

Drk interactive shell

Functions

completion 🔒

Auxiliary function to define the interactive shell completions.

handle_alias 🔒

Auxiliary function to define the alias command handling.

handle_alias_add 🔒

Auxiliary function to define the alias add subcommand handling.

handle_alias_remove 🔒

Auxiliary function to define the alias remove subcommand handling.

handle_alias_show 🔒

Auxiliary function to define the alias show subcommand handling.

handle_attach_fee 🔒

Auxiliary function to define the attach fee command handling.

handle_broadcast 🔒

Auxiliary function to define the broadcast command handling.

handle_completions 🔒

Auxiliary function to define the completions command handling.

handle_contract 🔒

Auxiliary function to define the contract command handling.

handle_contract_deploy 🔒

Auxiliary function to define the contract deploy subcommand handling.

handle_contract_export_data 🔒

Auxiliary function to define the contract export data subcommand handling.

handle_contract_generate_deploy 🔒

Auxiliary function to define the contract generate deploy subcommand handling.

handle_contract_list 🔒

Auxiliary function to define the contract list subcommand handling.

handle_contract_lock 🔒

Auxiliary function to define the contract lock subcommand handling.

handle_dao 🔒

Auxiliary function to define the dao command handling.

handle_dao_balance 🔒

Auxiliary function to define the dao balance subcommand handling.

handle_dao_create 🔒

Auxiliary function to define the dao create subcommand handling.

handle_dao_exec 🔒

Auxiliary function to define the dao exec subcommand handling.

handle_dao_import 🔒

Auxiliary function to define the dao import subcommand handling.

handle_dao_list 🔒

Auxiliary function to define the dao list subcommand handling.

handle_dao_mint 🔒

Auxiliary function to define the dao mint subcommand handling.

handle_dao_proposal 🔒

Auxiliary function to define the dao proposal subcommand handling.

handle_dao_proposal_import 🔒

Auxiliary function to define the dao proposal import subcommand handling.

handle_dao_proposals 🔒

Auxiliary function to define the dao proposals subcommand handling.

handle_dao_propose_generic 🔒

Auxiliary function to define the dao propose generic subcommand handling.

handle_dao_propose_transfer 🔒

Auxiliary function to define the dao propose transfer subcommand handling.

handle_dao_spend_hook 🔒

Auxiliary function to define the dao spent hook subcommand handling.

handle_dao_view 🔒

Auxiliary function to define the dao view subcommand handling.

handle_dao_vote 🔒

Auxiliary function to define the dao vote subcommand handling.

handle_explorer 🔒

Auxiliary function to define the explorer command handling.

handle_explorer_clear_reverted 🔒

Auxiliary function to define the explorer clear reverted subcommand handling.

handle_explorer_fetch_tx 🔒

Auxiliary function to define the explorer fetch transaction subcommand handling.

handle_explorer_scanned_blocks 🔒

Auxiliary function to define the explorer scanned blocks subcommand handling.

handle_explorer_simulate_tx 🔒

Auxiliary function to define the explorer simulate transaction subcommand handling.

handle_explorer_txs_history 🔒

Auxiliary function to define the explorer transactions history subcommand handling.

handle_inspect 🔒

Auxiliary function to define the inspect command handling.

handle_otc 🔒

Auxiliary function to define the otc command handling.

handle_otc_init 🔒

Auxiliary function to define the otc init subcommand handling.

handle_otc_inspect 🔒

Auxiliary function to define the otc inspect subcommand handling.

handle_otc_join 🔒

Auxiliary function to define the otc join subcommand handling.

handle_otc_sign 🔒

Auxiliary function to define the otc sign subcommand handling.

handle_ping 🔒

Auxiliary function to define the ping command handling.

handle_scan 🔒

Auxiliary function to define the scan command handling.

handle_spend 🔒

Auxiliary function to define the spend command handling.

handle_subscribe 🔒

Auxiliary function to define the subscribe command handling.

handle_token 🔒

Auxiliary function to define the token command handling.

handle_token_freeze 🔒

Auxiliary function to define the token freeze subcommand handling.

handle_token_generate_mint 🔒

Auxiliary function to define the token generate mint subcommand handling.

handle_token_import 🔒

Auxiliary function to define the token import subcommand handling.

handle_token_list 🔒

Auxiliary function to define the token list subcommand handling.

handle_token_mint 🔒

Auxiliary function to define the token mint subcommand handling.

handle_transfer 🔒

Auxiliary function to define the transfer command handling.

handle_unspend 🔒

Auxiliary function to define the unspend command handling.

handle_unsubscribe 🔒

Auxiliary function to define the unsubscribe command handling.

handle_wallet 🔒

Auxiliary function to define the wallet command handling.

handle_wallet_address 🔒

Auxiliary function to define the wallet address subcommand handling.

handle_wallet_addresses 🔒

Auxiliary function to define the wallet addresses subcommand handling.

handle_wallet_balance 🔒

Auxiliary function to define the wallet balance subcommand handling.

handle_wallet_coins 🔒

Auxiliary function to define the wallet coins subcommand handling.

handle_wallet_default_address 🔒

Auxiliary function to define the wallet default address subcommand handling.

handle_wallet_import_secrets 🔒

Auxiliary function to define the wallet import secrets subcommand handling.

handle_wallet_initialize 🔒

Auxiliary function to define the wallet initialize subcommand handling.

handle_wallet_keygen 🔒

Auxiliary function to define the wallet keygen subcommand handling.

handle_wallet_secrets 🔒

Auxiliary function to define the wallet secrets subcommand handling.

handle_wallet_tree 🔒

Auxiliary function to define the wallet tree subcommand handling.

help 🔒

Auxiliary function to print the help message.

hints 🔒

Auxiliary function to define the interactive shell hints.

interactive

Auxiliary function to start provided Drk as an interactive shell. Only sane/linenoise terminals are suported.

listen_for_line 🔒

Auxiliary function to listen for linenoise input line and handle background task messages.