Struct drk::Drk

pub struct Drk {
    pub wallet: Arc<WalletDb>,
    pub rpc_client: Option<RpcClient>,
}

CLI-util structure

Fields
wallet: Arc<WalletDb>

Wallet database operations handler

rpc_client: Option<RpcClient>

JSON-RPC client to execute requests to darkfid daemon

Implementations

impl Drk

pub async fn subscribe_blocks( &self, endpoint: Url, ex: Arc<Executor<'static>> ) -> Result<()>

Subscribes to darkfid’s JSON-RPC notification endpoint that serves new finalized blocks. Upon receiving them, all the transactions are scanned and we check if any of them call the money contract, and if the payments are intended for us. If so, we decrypt them and append the metadata to our wallet.

async fn scan_block(&self, block: &BlockInfo) -> Result<()>

scan_block will go over over transactions in a block and handle their calls based on the called contract. Additionally, will update last_scanned_block to the probided block height.

pub async fn scan_blocks(&self, reset: bool) -> Result<(), WalletDbError>

Scans the blockchain starting from the last scanned block, for relevant money transfer transactions. If reset flag is provided, Merkle tree state and coins are reset, and start scanning from beginning. Alternatively, it looks for a checkpoint in the wallet to reset and start scanning from.

async fn get_block_by_height(&self, height: u32) -> Result<BlockInfo>

pub async fn broadcast_tx(&self, tx: &Transaction) -> Result<String>

Broadcast a given transaction to darkfid and forward onto the network. Returns the transaction ID upon success.

pub async fn get_tx( &self, tx_hash: &TransactionHash ) -> Result<Option<Transaction>>

Queries darkfid for a tx with given hash.

pub async fn simulate_tx(&self, tx: &Transaction) -> Result<bool>

Simulate the transaction with the state machine.

pub async fn lookup_zkas( &self, contract_id: &ContractId ) -> Result<Vec<(String, Vec<u8>)>>

Try to fetch zkas bincodes for the given ContractId.

pub async fn get_tx_gas( &self, tx: &Transaction, include_fee: bool ) -> Result<u64>

Queries darkfid for given transaction’s gas.

impl Drk

pub async fn transfer( &self, amount: &str, token_id: TokenId, recipient: PublicKey ) -> Result<Transaction>

Create a payment transaction. Returns the transaction object on success.

impl Drk

pub async fn init_swap( &self, value_send: u64, token_send: TokenId, value_recv: u64, token_recv: TokenId ) -> Result<PartialSwapData>

Initialize the first half of an atomic swap

pub async fn join_swap(&self, partial: PartialSwapData) -> Result<Transaction>

Create a full transaction by inspecting and verifying given partial swap data, making the other half, and joining all this into a Transaction object.

pub async fn inspect_swap(&self, bytes: Vec<u8>) -> Result<()>

Inspect and verify a given swap (half or full) transaction

pub async fn sign_swap(&self, tx: &mut Transaction) -> Result<()>

Sign a given transaction by retrieving the secret key from the encrypted note and prepending it to the transaction’s signatures.

impl Drk

fn derive_token_attributes( &self, mint_authority: SecretKey, token_blind: BaseBlind ) -> TokenAttributes

Auxiliary function to derive TokenAttributes for provided secret key and token blind.

pub async fn import_mint_authority( &self, mint_authority: SecretKey, token_blind: BaseBlind ) -> Result<TokenId>

Import a token mint authority into the wallet.

pub async fn get_mint_authorities( &self ) -> Result<Vec<(TokenId, SecretKey, BaseBlind, bool)>>

pub async fn mint_token( &self, amount: &str, recipient: PublicKey, token_attrs: TokenAttributes ) -> Result<Transaction>

Create a token mint transaction. Returns the transaction object on success.

pub async fn freeze_token( &self, token_attrs: TokenAttributes ) -> Result<Transaction>

Create a token freeze transaction. Returns the transaction object on success.

impl Drk

pub async fn initialize_money(&self) -> Result<(), WalletDbError>

Initialize wallet with tables for the Money contract.

pub async fn money_keygen(&self) -> Result<(), WalletDbError>

Generate a new keypair and place it into the wallet.

pub async fn default_secret(&self) -> Result<SecretKey>

Fetch default secret key from the wallet.

pub async fn default_address(&self) -> Result<PublicKey>

Fetch default pubkey from the wallet.

pub async fn set_default_address(&self, idx: usize) -> Result<(), WalletDbError>

Set provided index address as default in the wallet.

pub async fn addresses(&self) -> Result<Vec<(u64, PublicKey, SecretKey, u64)>>

Fetch all pukeys from the wallet.

pub async fn get_money_secrets(&self) -> Result<Vec<SecretKey>>

Fetch all secret keys from the wallet.

pub async fn import_money_secrets( &self, secrets: Vec<SecretKey> ) -> Result<Vec<PublicKey>>

Import given secret keys into the wallet. If the key already exists, it will be skipped. Returns the respective PublicKey objects for the imported keys.

pub async fn money_balance(&self) -> Result<HashMap<String, u64>>

Fetch known unspent balances from the wallet and return them as a hashmap.

pub async fn get_coins( &self, fetch_spent: bool ) -> Result<Vec<(OwnCoin, bool, String)>>

Fetch all coins and their metadata related to the Money contract from the wallet. Optionally also fetch spent ones. The boolean in the returned tuple notes if the coin was marked as spent.

pub async fn get_token_coins(&self, token_id: &TokenId) -> Result<Vec<OwnCoin>>

Fetch provided token unspend balances from the wallet.

async fn parse_coin_record( &self, row: &[Value] ) -> Result<(OwnCoin, bool, String)>

Auxiliary function to parse a MONEY_COINS_TABLE record. The boolean in the returned tuple notes if the coin was marked as spent.

pub async fn add_alias( &self, alias: String, token_id: TokenId ) -> Result<(), WalletDbError>

Create an alias record for provided Token ID.

pub async fn get_aliases( &self, alias_filter: Option<String>, token_id_filter: Option<TokenId> ) -> Result<HashMap<String, TokenId>>

Fetch all aliases from the wallet. Optionally filter using alias name and/or token id.

pub async fn get_aliases_mapped_by_token( &self ) -> Result<HashMap<String, String>>

Fetch all aliases from the wallet, mapped by token id.

pub async fn remove_alias(&self, alias: String) -> Result<(), WalletDbError>

Remove provided alias record from the wallet database.

pub async fn unspend_coin(&self, coin: &Coin) -> Result<(), WalletDbError>

Mark a given coin in the wallet as unspent.

pub async fn put_money_tree( &self, tree: &MerkleTree ) -> Result<(), WalletDbError>

Replace the Money Merkle tree in the wallet.

pub async fn get_money_tree(&self) -> Result<MerkleTree>

Fetch the Money Merkle tree from the wallet.

pub async fn last_scanned_block(&self) -> Result<u32, WalletDbError>

Get the last scanned block height from the wallet.

async fn parse_call_data( &self, data: &[u8] ) -> Result<(Vec<Nullifier>, Vec<Coin>, Vec<AeadEncryptedNote>, Vec<TokenId>)>

Auxiliary function to grab all the nullifiers, coins, notes and freezes from transaction data.

pub async fn apply_tx_money_data( &self, data: &[u8], tx_hash: &String ) -> Result<()>

Append data related to Money contract transactions into the wallet database.

pub async fn mark_tx_spend(&self, tx: &Transaction) -> Result<()>

Mark provided transaction input coins as spent.

pub async fn mark_spent_coin( &self, coin: &Coin, spent_tx_hash: &String ) -> Result<(), WalletDbError>

Mark a coin in the wallet as spent

pub async fn mark_spent_coins( &self, nullifiers: &[Nullifier], spent_tx_hash: &String ) -> Result<()>

Marks all coins in the wallet as spent, if their nullifier is in the given set

pub async fn reset_money_tree(&self) -> Result<(), WalletDbError>

Reset the Money Merkle tree in the wallet

pub async fn reset_money_coins(&self) -> Result<(), WalletDbError>

Reset the Money coins in the wallet

pub async fn get_token(&self, input: String) -> Result<TokenId>

Retrieve token by provided string. Input string represents either an alias or a token id.

pub async fn append_fee_call( &self, tx: &Transaction, public_key: PublicKey, money_merkle_tree: &MerkleTree, fee_pk: &ProvingKey, fee_zkbin: &ZkBinary, spent_coins: Option<&[OwnCoin]> ) -> Result<(ContractCall, Vec<Proof>, Vec<SecretKey>)>

Create and append a Money::Fee call to a given [Transaction].

Optionally takes a set of spent coins in order not to reuse them here.

Returns the Fee call, and all necessary data and parameters related.

impl Drk

pub async fn initialize_dao(&self) -> Result<(), WalletDbError>

Initialize wallet with tables for the DAO contract.

pub async fn put_dao_trees( &self, daos_tree: &MerkleTree, proposals_tree: &MerkleTree ) -> Result<(), WalletDbError>

Replace the DAO Merkle trees in the wallet.

pub async fn get_dao_trees(&self) -> Result<(MerkleTree, MerkleTree)>

Fetch DAO Merkle trees from the wallet.

pub async fn get_dao_secrets(&self) -> Result<Vec<SecretKey>>

Fetch all DAO secret keys from the wallet.

pub async fn get_daos(&self) -> Result<Vec<Dao>>

Fetch all known DAOs from the wallet.

async fn parse_dao_proposal( &self, dao: &Dao, row: &[Value] ) -> Result<DaoProposal>

Auxiliary function to parse a proposal record row.

pub async fn get_dao_proposals(&self, dao_id: u64) -> Result<Vec<DaoProposal>>

Fetch all known DAO proposals from the wallet given a DAO ID.

pub async fn apply_tx_dao_data( &self, data: &[u8], tx_hash: TransactionHash, call_idx: u8, confirm: bool ) -> Result<()>

Append data related to DAO contract transactions into the wallet database. Optionally, if confirm is true, also append the data in the Merkle trees, etc.

pub async fn confirm_daos(&self, daos: &[Dao]) -> Result<(), WalletDbError>

Confirm already imported DAO metadata into the wallet. Here we just write the leaf position, tx hash, and call index. Panics if the fields are None.

pub async fn unconfirm_daos(&self, daos: &[Dao]) -> Result<(), WalletDbError>

Unconfirm imported DAOs by removing the leaf position, txid, and call index.

pub async fn put_dao_proposals(&self, proposals: &[DaoProposal]) -> Result<()>

Import given DAO proposals into the wallet.

pub async fn put_dao_votes( &self, votes: &[DaoVote] ) -> Result<(), WalletDbError>

Import given DAO votes into the wallet.

pub async fn reset_dao_trees(&self) -> Result<(), WalletDbError>

Reset the DAO Merkle trees in the wallet.

pub async fn reset_daos(&self) -> Result<(), WalletDbError>

Reset confirmed DAOs in the wallet.

pub async fn reset_dao_proposals(&self) -> Result<(), WalletDbError>

Reset all DAO proposals in the wallet.

pub async fn reset_dao_votes(&self) -> Result<(), WalletDbError>

Reset all DAO votes in the wallet.

pub async fn import_dao( &self, dao_name: String, dao_params: DaoParams ) -> Result<()>

Import given DAO params into the wallet with a given name.

pub async fn get_dao_id_by_alias(&self, alias_filter: &str) -> Result<u64>

Retrieve DAO ID using provided alias filter.

pub async fn get_dao_id(&self, alias: &str) -> Result<u64>

Convenience function. Interprets the alias either as the DAO alias or its ID.

pub async fn get_dao_by_id(&self, dao_id: u64) -> Result<Dao>

Fetch a DAO given a numeric ID.

pub async fn dao_list(&self, dao_id: Option<u64>) -> Result<()>

List DAO(s) imported in the wallet. If an ID is given, just print the metadata for that specific one, if found.

async fn dao_list_single(&self, dao_id: u64) -> Result<()>

Retrieve DAO for provided ID and print its metadata.

pub async fn dao_balance(&self, dao_id: u64) -> Result<HashMap<String, u64>>

Fetch known unspent balances from the wallet for the given DAO ID

pub async fn get_dao_proposal_by_id( &self, proposal_id: u64 ) -> Result<DaoProposal>

Fetch a DAO proposal by its ID

pub async fn get_dao_proposal_votes( &self, proposal_id: u64 ) -> Result<Vec<DaoVote>>

pub async fn dao_mint(&self, dao_id: u64) -> Result<Transaction>

Mint a DAO on-chain

pub async fn dao_propose( &self, dao_id: u64, _recipient: PublicKey, amount: u64, token_id: TokenId ) -> Result<Transaction>

Create a DAO proposal

pub async fn dao_vote( &self, dao_id: u64, proposal_id: u64, vote_option: bool, weight: u64 ) -> Result<Transaction>

Vote on a DAO proposal

pub async fn dao_exec( &self, _dao: Dao, _proposal: DaoProposal ) -> Result<Transaction>

Import given DAO votes into the wallet This function is really bad but I’m also really tired and annoyed.

impl Drk

pub async fn initialize_deployooor(&self) -> Result<(), WalletDbError>

Initialize wallet with tables for the Deployooor contract.

pub async fn deploy_auth_keygen(&self) -> Result<(), WalletDbError>

Generate a new deploy authority keypair and place it into the wallet

pub async fn list_deploy_auth(&self) -> Result<Vec<(i64, ContractId, bool)>>

List contract deploy authorities from the wallet

async fn get_deploy_auth(&self, idx: u64) -> Result<Keypair>

Retrieve a deploy authority keypair given an index

pub async fn deploy_contract( &self, deploy_auth: u64, wasm_bincode: Vec<u8>, deploy_ix: Vec<u8> ) -> Result<Transaction>

Create a contract deployment transaction

pub async fn lock_contract(&self, deploy_auth: u64) -> Result<Transaction>

Create a contract redeployment lock transaction

impl Drk

pub async fn insert_tx_history_record( &self, tx: &Transaction ) -> Result<String, WalletDbError>

Insert a Transaction history record into the wallet.

pub async fn insert_tx_history_records( &self, txs: &[Transaction] ) -> Result<Vec<String>, WalletDbError>

Insert a slice of [Transaction] history records into the wallet.

pub async fn get_tx_history_record( &self, tx_hash: &str ) -> Result<(String, String, Transaction)>

Get a transaction history record.

pub async fn get_txs_history( &self ) -> Result<Vec<(String, String)>, WalletDbError>

Fetch all transactions history records, excluding bytes column.

pub async fn update_tx_history_records_status( &self, txs_hashes: &[String], status: &str ) -> Result<(), WalletDbError>

Update given transactions history record statuses to the given one.

pub async fn update_all_tx_history_records_status( &self, status: &str ) -> Result<(), WalletDbError>

Update all transaction history records statuses to the given one.

impl Drk

pub(crate) async fn new( wallet_path: String, wallet_pass: String, endpoint: Option<Url>, ex: Arc<Executor<'static>> ) -> Result<Self>

pub(crate) async fn initialize_wallet(&self) -> Result<()>

Initialize wallet with tables for drk

pub(crate) async fn ping(&self) -> Result<()>

Auxiliary function to ping configured darkfid daemon for liveness.